Move 'atMostOne' utils to common package
