+++ backend.native/compiler/ir/backend.native/src/org/jetbrains/kotlin/backend/common/Utils.kt
@@ -48,3 +48,17 @@ fun <E> MutableList<E>.pop() = this.removeAt(size - 1)
 
 fun <E> MutableList<E>.peek(): E? = if (size == 0) null else this[size - 1]
 
+fun <T> Collection<T>.atMostOne(): T? {
+    return when (this.size) {
+        0 -> null
+        1 -> this.iterator().next()
+        else -> throw IllegalArgumentException("Collection has more than one element.")
+    }
+}
+
+inline fun <T> Iterable<T>.atMostOne(predicate: (T) -> Boolean): T? = this.filter(predicate).atMostOne()
+
+fun <T: Any> T.onlyIf(condition: T.()->Boolean, then: (T)->Unit): T {
+    if (this.condition()) then(this)
+    return this
+}
